Quadrant Selection
Your job is to take a point and determine the quadrant it is in. You can assume that neither of the two coordinates will be 0.

Input:
The first line of input contains the integer x (−1000 ≤ x ≤ 1000; x ≠ 0). The second line of input contains the integer y (−1000 ≤ y ≤ 1000; y ≠ 0).

Output:
Output the quadrant number (1, 2, 3 or 4) for the point (x,y).