Sibice
Help Mirko determine which of the matches fit in the box his mom gave him. A match fits in the box if its entire length can lie on the bottom of the box. Mirko examines the matches one by one.

Input
The first line of input contains an integer N (1 ≤ N ≤ 50), the number of matches on the floor, and two integers W and H, the dimensions of the box (1 ≤ W ≤ 100, 1 ≤ H ≤ 100).

Each of the following N lines contains a single integer between 1 and 1000 (inclusive), the length of one match.

Output
For each match, in the order they were given in the input, output on a separate line “DA” if the match fits in the box or “NE” if it does not.