Faktor
Input
First and only line of input will contain 2 integers, A (1 ≤ A ≤ 100), the number of articles you plan to publish and I (1 ≤ I ≤ 100), the impact factor the owners require.

Output
The first and only line of output should contain one integer, the minimal number of scientists you need to bribe.